backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
spotipy
python-multipart
ffmpeg-python
requests
uvloop
httptools